removed unnecessary includes and put threecenter.cc files into one file
